Dataset class, render radar images
I've managed to download all data, extract it, and make a basic Dataset class that has some methods to access the data. I then used cartopy (next-gen Basemap apparently) to plot it on a map, where I had to fiddle a bit with the coordinate systems but now everything worked out fine. Quite impressive to see it visualized in this way.