Map-reduce average all data
I implemented a simple map-reduce kind of operation, where there are workers that extract the image array from the h5 files and then sum the results. In the end all data is averaged to produce a nice visualization of the radar's specifics, like where it is more and less sensitive. Takes close to 20min on the g2.2xlarge instance I'm running it on currently.