Convert to mm/h, transforms, and metrics
Did some more refactoring, fixing, and implemented conversion to mm/h, and some transforms. Plotting function can plot a video of a time series. But the best part is that the metrics from the paper are implemented, and working on a comparison of real data with knmi predicted data. Next step is to implement the LSTM and start training.