Treshold research, synthetic dataset
I was curious what kind of influence the treshold has on the performance metrics. Turns out the treshold can be set a little lower for better scores because a lot of the rain is quite little, less than .5mm/h like in the paper. See plots. Also made a synthetic moving ball dataset that can be used to train the LSTM.