Training ConvLSTM on synthetic dataset
Does not really work yet. Trained it with one layer and one frame first, and discovered that the output range is kind of funky, not what you'd expect. Would have to do some more research but for now it suffices to use the [-.5,.5] output range of the LSTM (theoretically it can go from -1 to 1 due to the tanh). Now implemented an encoder-decoder architecture but have yet to get it to work properly.