Basics in place, unittests half coverage
I implemented basically everything in the VPN, which is MU, RMB, ConvLSTM, and combined these to form an Encoder and a Decoder. The MU and RMB have simple test coverage, next step is to test the other modules before starting to use them.

Also added a README with a discussion of some basic design choices and open problems.

The ConvLSTM implementation now lives as a local file, which is more convenient than the submodule.