All basic tests passing (ConvLSTM, Encoder, Decoder + inference)
These tests are simple but I realized that they don't have to be very eleborate to be useful. Some of the functionality I'm still defining so it's not a problem that there are no checks for all details. The test for all modules, including decoder inference, are passing. Next step is to integrate in a top-level model class and set up a training loop and see if it actually trains.