Clean up loss func
As long as the tensors have the same number of elements ordered in the same way, they can be used with binary cross-entropy loss. So we don't have to loop but can just calculate it once for the predictions and targets we want. Also, the number of pixel values in the ball dataset are just 2.